+++ source/renderer/app/utils/colors.js
@@ -1,43 +1,38 @@
 // @flow
-import { rangeMap } from './rangeMap';
+import chroma from 'chroma-js';
 
 // Ranking 001: hsla(142, 76%, 45%, 1)
 // Ranking 100: hsla(15, 97%, 58%, 1)
 
-const RANKING_COLORS_RANGE = {
-  hue: [142, 15],
-  saturation: [76, 97],
-  lighness: [45, 58],
+type RangeOptions = {
+  colors?: Array<any>,
+  domain?: Array<number>,
+  darken?: number,
+  brighten?: number,
+  alpha?: number,
+  reverse?: boolean,
 };
 
-const getHSLParam = (ranking: number, param: string, offset?: number) =>
-  rangeMap(
-    ranking,
-    1,
-    100,
-    RANKING_COLORS_RANGE[param][0],
-    RANKING_COLORS_RANGE[param][1]
-  ) + offset;
-
-type Options = {
-  hueOffset?: number,
-  saturationOffset?: number,
-  lighnessOffset?: number,
+const defaultRangeOptions = {
+  colors: ['#1cca5b', '#fc602c'],
+  domain: [0, 99],
+  darken: 0,
+  brighten: 0,
+  alpha: 1,
+  reverse: false,
 };
 
-const defaultOptions = {
-  hueOffset: 0,
-  saturationOffset: 0,
-  lighnessOffset: 0,
-};
-
-export const getHSLColor = (ranking: number, options?: Options = {}) => {
-  const { hueOffset, saturationOffset, lighnessOffset } = {
-    ...defaultOptions,
+export const getColorFromRange = (index: number, options?: RangeOptions) => {
+  const { colors, domain: originalDomain, darken, brighten, alpha, reverse } = {
+    ...defaultRangeOptions,
     ...options,
   };
-  const hue = getHSLParam(ranking, 'hue', hueOffset);
-  const saturation = getHSLParam(ranking, 'saturation', saturationOffset);
-  const lighness = getHSLParam(ranking, 'lighness', lighnessOffset);
-  return `hsl(${hue}, ${saturation}%, ${lighness}%)`;
+  const domain = originalDomain.slice();
+  if (reverse) domain.reverse();
+  const scale = chroma.scale(colors).domain(domain);
+  return scale(index)
+    .darken(darken)
+    .brighten(brighten)
+    .alpha(alpha)
+    .hex();
 };