+++ source/renderer/app/themes/daedalus/dark-blue.js
@@ -326,6 +326,17 @@ const DARK_BLUE_THEME_CONFIG = {
     '--theme-loading-status-icons-tooltip-color': '#4b5a68',
     '--theme-loading-spinner-color': '#e9f4fe',
   },
+  manualUpdate: {
+    '--theme-manual-update-overlay-background-color': 'rgba(38, 51, 69, 0.96)',
+    '--theme-manual-update-overlay-button-background-color':
+      'rgba(0, 0, 0, 0.1)',
+    '--theme-manual-update-overlay-button-background-color-hover': '#fafbfc',
+    '--theme-manual-update-overlay-button-text-color-hover': '#263345',
+    '--theme-manual-update-overlay-button-border-color': '#fafbfc',
+    '--theme-manual-update-overlay-text-color': 'rgba(250, 251, 252, 0.7)',
+    '--theme-manual-update-overlay-text-highlight-color': '#fafbfc',
+    '--theme-manual-update-overlay-title-text-color': '#fafbfc',
+  },
   mnemonic: {
     '--theme-backup-mnemonic-background-color': 'rgba(233, 244, 254, 0.05)',
     '--theme-mnemonic-background-color': 'rgba(83, 99, 112, 0.3)',
@@ -478,8 +489,8 @@ const DARK_BLUE_THEME_CONFIG = {
     '--theme-staking-stake-pool-tooltip-link-color': '#2ab467',
     '--theme-staking-stake-pool-tooltip-table-title-color': '#e9f4fe',
     '--theme-staking-stake-pool-tooltip-table-param-color': '#e9f4fe',
-    '--theme-staking-stake-pool-tooltip-table-param-retirement-background-color':
-      'rgba(234, 76, 91, 0.3)',
+    '--theme-staking-stake-pool-tooltip-retirement-text-color': '#fafbfc',
+    '--theme-staking-stake-pool-tooltip-retirement-background-color': '#ea4c5b',
     '--theme-staking-stake-pool-tooltip-delegate-button-background-color':
       'rgba(0, 0, 0, 0.1)',
     '--theme-staking-stake-pool-tooltip-delegate-button-hover-background-color':